+++ lib/Extension/Prophecy/Tests/Integration/WorseReflection/ProphecyMemberContextResolverTest.php
@@ -28,9 +28,37 @@ class ProphecyMemberContextResolverTest extends IntegrationTestCase
                 }
 
                 $prophet = (new Foobar())->prophesize(Hello::class);
+
                 wrAssertType('Prophecy\Prophecy\ObjectProphecy<Hello>', $prophet);
-                wrAssertType('Prophecy\Prophecy\MethodProphecy<string>', $prophet->bar());
+                wrAssertType('Prophecy\Prophecy\MethodProphecy<Hello>', $prophet->bar());
+                wrAssertType('Prophecy\Prophecy\MethodProphecy<Hello>', $prophet->bar()->will());
+                wrAssertType('Prophecy\Prophecy\ObjectProphecy<Hello>', $prophet->bar()->getObjectProphecy());
+                wrAssertType('Prophecy\Prophecy\MethodProphecy<Hello>', $prophet->bar()->getObjectProphecy()->bar());
+                wrAssertType('Prophecy\Prophecy\MethodProphecy<Hello>', $prophet->bar()->willReturn('')->getObjectProphecy()->bar());
+                wrAssertType('string', $prophet->bar()->getMethodName());
                 wrAssertType('Hello', $prophet->reveal());
+                EOT
+            ,
+        );
+    }
+    public function testMethodProphecy(): void
+    {
+        $this->resolve(
+            <<<'EOT'
+                <?php
+                class Hello {
+                    public function bar(): string
+                    {
+                    }
+                }
+                class Foobar {
+                    public function prophesize(string $class): \Prophecy\Prophecy\ObjectProphecy {}
+                }
+                $prophet = (new Foobar())->prophesize(Hello::class);
+
+                wrAssertType('Prophecy\Prophecy\MethodProphecy<Hello>', $prophet->bar()->willReturn(''));
+                wrAssertType('Prophecy\Prophecy\ObjectProphecy<Hello>', $prophet->bar()->willReturn('')->getObjectProphecy());
+                wrAssertType('Prophecy\Prophecy\MethodProphecy<Hello>', $prophet->bar()->willReturn('')->getObjectProphecy()->bar());
                 EOT
             ,
         );
@@ -57,7 +85,7 @@ class ProphecyMemberContextResolverTest extends IntegrationTestCase
                     public function hello(): void
                     {
                         wrAssertType('Prophecy\Prophecy\ObjectProphecy<Hello>', $this->hello);
-                        wrAssertType('Prophecy\Prophecy\MethodProphecy<string>', $this->hello->bar());
+                        wrAssertType('Prophecy\Prophecy\MethodProphecy<Hello>', $this->hello->bar());
                     }
                 }
                 EOT